SPA beginning.
This step includes a working proof of concept for my approach for a single page application, with a working navbar for desktop and mobile that successfully replaces the content on screen with the requested "page" of content, and a successful browser history and address handling script. CSS for about page is NOT implemented in this commit.